+++ app/src/main/java/com/studentlifemanager/ui/MainActivity.kt
@@ -1,23 +1,12 @@
 package com.studentlifemanager.ui
 
 import android.os.Bundle
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.findNavController
 import androidx.navigation.ui.setupWithNavController
 import com.studentlifemanager.R
-import com.studentlifemanager.common.utils.Constant.DATE_PICKER_TAG
-import com.studentlifemanager.common.utils.Constant.FULL_MONTH
-import com.studentlifemanager.common.utils.Constant.FULL_YEAR
-import com.studentlifemanager.common.utils.MainViewModel
 import com.studentlifemanager.databinding.ActivityMainBinding
-import com.studentlifemanager.expense.utils.IViewCallback
-import com.studentlifemanager.expense.utils.MonthYearPickerDialog
 import dagger.hilt.android.AndroidEntryPoint
-import java.text.DateFormat
-import java.text.SimpleDateFormat
-import java.time.ZoneId
-import java.util.Date
 
 /**
  * This is the main activity of the app
@@ -28,10 +17,9 @@ import java.util.Date
  * @version 2.0
  */
 @AndroidEntryPoint
-class MainActivity : AppCompatActivity(), IViewCallback {
+class MainActivity : AppCompatActivity() {
 
     private val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
-    private val mainViewModel: MainViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -43,54 +31,5 @@ class MainActivity : AppCompatActivity(), IViewCallback {
         binding.navView.setupWithNavController(
             findNavController(R.id.nav_host_fragment_activity_main)
         )
-
-        // bing variables
-        binding.callback = this
-
-        // below toolbar, date and year is visible by default we are showing current year nad month
-        // this function is used for this purpose
-        setDefaultDate()
-
     }
-
-    /**
-     * This is call back function for view click in order to select month or year
-     *
-     */
-    override fun viewCallBack() {
-        selectMonth()
-    }
-
-
-    /**
-     * When user click on month or year this function will open a date picker dialog
-     * Here user can select year or month in order to get respective expenses
-     */
-    private fun selectMonth() {
-        MonthYearPickerDialog(Date()).apply {
-            setListener { view, year, month, dayOfMonth ->
-                mainViewModel.selectItem(month + 1)
-                binding.tvMonth.text = getMonth()[month]
-                binding.tvYear.text = year.toString()
-            }
-            show(supportFragmentManager, DATE_PICKER_TAG)
-        }
-    }
-
-    /**
-     * below toolbar, date and year is visible
-     * by default we are showing current year nad month
-     *
-     * this function is used for this purpose
-     */
-    private fun setDefaultDate() {
-        val monthFormat: DateFormat = SimpleDateFormat(FULL_MONTH)
-        val yearFormat: DateFormat = SimpleDateFormat(FULL_YEAR)
-        val date = Date()
-        val localDate = date.toInstant().atZone(ZoneId.systemDefault()).toLocalDate()
-        binding.tvMonth.text = monthFormat.format(date)
-        binding.tvYear.text = yearFormat.format(date)
-        mainViewModel.selectItem(localDate.monthValue)
-    }
-
 }
