+++ app/src/main/res/layout/activity_main.xml
@@ -5,10 +5,6 @@
 
     <data>
 
-        <variable
-            name="callback"
-            type="com.studentlifemanager.expense.utils.IViewCallback" />
-
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -21,70 +17,13 @@
             android:layout_width="match_parent"
             android:layout_height="?actionBarSize"
             android:background="@color/colorPrimary"
-            app:layout_constraintBottom_toTopOf="@+id/layDate"
+            app:layout_constraintBottom_toTopOf="@+id/nav_host_fragment_activity_main"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
             app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
             app:titleTextAppearance="@style/Toolbar.TitleText"
             app:titleTextColor="@color/white" />
-
-        <androidx.appcompat.widget.LinearLayoutCompat
-            android:id="@+id/layDate"
-            android:layout_width="match_parent"
-            android:layout_height="45dp"
-            android:background="@color/colorPrimary"
-            android:gravity="center"
-            app:layout_constraintBottom_toTopOf="@+id/nav_host_fragment_activity_main"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/toolbar"
-            tools:ignore="MissingConstraints">
-
-            <ImageView
-                android:id="@+id/imgMothLeft"
-                android:layout_width="@dimen/dp_35"
-                android:layout_height="match_parent"
-                android:layout_marginEnd="@dimen/dimen_15"
-                android:onClick="@{()-> callback.viewCallBack()}"
-                android:padding="@dimen/dim_5dp"
-                android:src="@drawable/ic_left"
-                app:tint="@color/white" />
-
-            <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/tvMonth"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:fontFamily="@font/montserrat_alternates_regular"
-                android:gravity="center"
-                android:onClick="@{()-> callback.viewCallBack()}"
-                android:textColor="@color/white"
-                android:textSize="@dimen/dim_17sp"
-                tools:text="January," />
-
-            <androidx.appcompat.widget.AppCompatTextView
-                android:id="@+id/tvYear"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:layout_marginStart="@dimen/dim_5dp"
-                android:fontFamily="@font/montserrat_alternates_regular"
-                android:gravity="center"
-                android:onClick="@{()-> callback.viewCallBack()}"
-                android:text="2023"
-                android:textColor="@color/white"
-                android:textSize="@dimen/dim_17sp" />
-
-            <ImageView
-                android:id="@+id/imgMonthRight"
-                android:layout_width="@dimen/dp_35"
-                android:layout_height="match_parent"
-                android:layout_marginStart="@dimen/dimen_15"
-                android:onClick="@{()-> callback.viewCallBack()}"
-                android:padding="@dimen/dim_5dp"
-                android:src="@drawable/ic_right"
-                app:tint="@color/white" />
-
-        </androidx.appcompat.widget.LinearLayoutCompat>
 
         <fragment
             android:id="@+id/nav_host_fragment_activity_main"
@@ -95,7 +34,7 @@
             app:layout_constraintBottom_toTopOf="@+id/nav_view"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/layDate"
+            app:layout_constraintTop_toBottomOf="@+id/toolbar"
             app:navGraph="@navigation/mobile_navigation" />
 
         <com.google.android.material.bottomnavigation.BottomNavigationView
